Code for testing ( and using ) coyote standalone.
The adapters are very simple, more as example. The goal is to have a
minimal http connector, and play with nio and other apis in a simpler 
environment

About MessageWriter and MessageReader - they are copies of
Catalina/connector/OutputBuffer and InputBuffer. Renamed to avoid
confusion. Long term, they should be part of o.a.coyote package, there
is no dep on catalina and are usefull in in connector. 



git-svn-id: https://svn.apache.org/repos/asf/tomcat/sandbox@331062 13f79535-47bb-0310-9956-ffa450edef68