Refactored modeler - using DynamicMbeans instead of model mbeans.
The name 'modeler' is still valid - it 'models' mbeans, but it doesn't have to use
the heavy ModelMBean interfaces. 

Most of the model-related features, and other things not used in tomcat are removed. 

Tested on tomcat6 - but since it's a big change I want to have it in sandbox first



git-svn-id: https://svn.apache.org/repos/asf/tomcat/sandbox@407661 13f79535-47bb-0310-9956-ffa450edef68